GET,POST処理をserver.rsに移動
